+++ maat-court-data-api/src/main/java/gov/uk/courtdata/common/dto/ErrorDTO.java
@@ -1,4 +1,4 @@
-package gov.uk.courtdata.dto;
+package gov.uk.courtdata.common.dto;
 
 
 import lombok.Builder;
