+++ maat-court-data-api/src/main/java/gov/uk/courtdata/common/repository/CrownCourtProcessingRepository.java
@@ -1,9 +1,7 @@
-package gov.uk.courtdata.repository;
+package gov.uk.courtdata.common.repository;
 
 import gov.uk.courtdata.entity.RepOrderCCOutComeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.jpa.repository.query.Procedure;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
