+++ maat-court-data-api/src/main/java/gov/uk/courtdata/crowncourt/model/UpdateSentenceOrder.java
@@ -1,4 +1,4 @@
-package gov.uk.courtdata.model;
+package gov.uk.courtdata.crowncourt.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
