+++ maat-court-data-api/src/main/java/gov/uk/courtdata/laastatus/controller/LaaStatusUpdateController.java
@@ -2,7 +2,7 @@ package gov.uk.courtdata.laastatus.controller;
 
 import com.amazonaws.xray.spring.aop.XRayEnabled;
 import com.google.gson.Gson;
-import gov.uk.courtdata.dto.ErrorDTO;
+import gov.uk.courtdata.common.dto.ErrorDTO;
 import gov.uk.courtdata.enums.LoggingData;
 import gov.uk.courtdata.enums.MessageType;
 import gov.uk.courtdata.exception.MAATCourtDataException;
