+++ maat-court-data-api/src/main/java/gov/uk/courtdata/link/controller/LinkController.java
@@ -1,7 +1,7 @@
 package gov.uk.courtdata.link.controller;
 
 import com.amazonaws.xray.spring.aop.XRayEnabled;
-import gov.uk.courtdata.dto.ErrorDTO;
+import gov.uk.courtdata.common.dto.ErrorDTO;
 import gov.uk.courtdata.enums.LoggingData;
 import gov.uk.courtdata.link.validator.PreConditionsValidator;
 import gov.uk.courtdata.model.CaseDetailsValidate;
