+++ maat-court-data-api/src/main/java/gov/uk/courtdata/offence/controller/OffenceController.java
@@ -1,7 +1,7 @@
 package gov.uk.courtdata.offence.controller;
 
 import com.amazonaws.xray.spring.aop.XRayEnabled;
-import gov.uk.courtdata.dto.ErrorDTO;
+import gov.uk.courtdata.common.dto.ErrorDTO;
 import gov.uk.courtdata.dto.OffenceDTO;
 import gov.uk.courtdata.offence.service.OffenceService;
 import io.swagger.v3.oas.annotations.Operation;
