+++ maat-court-data-api/src/main/java/gov/uk/courtdata/reporder/controller/RepOrderCapitalController.java
@@ -1,7 +1,7 @@
 package gov.uk.courtdata.reporder.controller;
 
 import com.amazonaws.xray.spring.aop.XRayEnabled;
-import gov.uk.courtdata.dto.ErrorDTO;
+import gov.uk.courtdata.common.dto.ErrorDTO;
 import gov.uk.courtdata.reporder.service.RepOrderCapitalService;
 import gov.uk.courtdata.validator.MaatIdValidator;
 import io.swagger.v3.oas.annotations.Operation;
