+++ maat-court-data-api/src/main/java/gov/uk/courtdata/wqoffence/controller/WQOffenceController.java
@@ -2,7 +2,7 @@ package gov.uk.courtdata.wqoffence.controller;
 
 
 import com.amazonaws.xray.spring.aop.XRayEnabled;
-import gov.uk.courtdata.dto.ErrorDTO;
+import gov.uk.courtdata.common.dto.ErrorDTO;
 import gov.uk.courtdata.wqoffence.service.WQOffenceService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
