+++ maat-court-data-api/src/test/java/gov/uk/courtdata/integration/repOrder/CCOutcomeControllerIntegrationTest.java
@@ -3,11 +3,11 @@ package gov.uk.courtdata.integration.repOrder;
 import gov.uk.MAATCourtDataApplication;
 import gov.uk.courtdata.builder.TestEntityDataBuilder;
 import gov.uk.courtdata.builder.TestModelDataBuilder;
+import gov.uk.courtdata.common.repository.CrownCourtProcessingRepository;
 import gov.uk.courtdata.entity.RepOrderCCOutComeEntity;
 import gov.uk.courtdata.entity.RepOrderEntity;
 import gov.uk.courtdata.integration.MockServicesConfig;
 import gov.uk.courtdata.model.RepOrderCCOutcome;
-import gov.uk.courtdata.repository.CrownCourtProcessingRepository;
 import gov.uk.courtdata.repository.RepOrderRepository;
 import gov.uk.courtdata.util.MockMvcIntegrationTest;
 import org.assertj.core.api.SoftAssertions;
