+++ maat-court-data-api/src/test/java/gov/uk/courtdata/util/MockMvcIntegrationTest.java
@@ -1,7 +1,7 @@
 package gov.uk.courtdata.util;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import gov.uk.courtdata.dto.ErrorDTO;
+import gov.uk.courtdata.common.dto.ErrorDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.http.HttpStatus;
